Make index.php cleaner

+++ index.php
@@ -49,26 +49,24 @@ try {
 
 
         // simple command for sending a message
-        if ($text === "/testBot") {
-            $chat->sendMessage("testBot ok", true);
-        }
+        if ($text === "/testBot") $chat->sendMessage("testBot ok", true);
 
 
         // command to delete the message in reply with json_payload enabled
         // "$replyToMessage !== null and $usernameReply !== null"  checks if is a reply
-        if ($text === "/deleteMessage" and $replyToMessage !== null and $usernameReply !== null) {
-            $replyToMessage->delete(null, true);
-        }
+        if ($text === "/deleteMessage" and $replyToMessage !== null and $usernameReply !== null) $replyToMessage->delete(null, true);
 
 
         // command to change chat's title (if is a group/supergroup)
         // example of usage: "/changeTitle i choose this title" the title of the chat will be "i choose this title"
-        if (strpos($text, "/changeTitle") === 0) {
+        if (strpos($text, "/changeTitle") === 0) 
+        {
             $ai = explode(" ", $text);
             unset($ai[0]);
             $ei = implode(" ", $ai);
 
-            $Bot->setChatTitle([
+            $Bot->setChatTitle
+            ([
                 "chat_id" => $chatid,
                 "title" => $ei
             ]);
@@ -76,37 +74,35 @@ try {
 
 
         // command to get Datacenter of user in reply
-        if ($text === "/getUserDc") {
-            $chat->sendMessage("@$usernameReply datacenter: " . $userReply->getDC());
-        }
+        if ($text === "/getUserDc") $chat->sendMessage("@$usernameReply datacenter: " . $userReply->getDC());
 
 
         // command to pin message in reply
         // "$replyToMessage !== null and $usernameReply !== null"  checks if is a reply
-        if (($text === "/pinMessage") and ($replyToMessage !== null and $usernameReply !== null)) {
-            $message->reply_to_message->pin();
-        }
+        if (($text === "/pinMessage") and ($replyToMessage !== null and $usernameReply !== null)) $message->reply_to_message->pin();
 
 
         // command to unpin the pinned message
-        if ($text === "/unpinMessage") {
-            $unpinChatMessage = $Bot->unpinChatMessage([
+        if ($text === "/unpinMessage") 
+        {
+            $Bot->unpinChatMessage
+            ([
                 "chat_id" => $chatid    // chat_id of the chat
             ]);
 
             // checks if there is a pinned message in the chat
-            if ($unpinChatMessage != true) {
-                $chat->sendMessage("there isn't any pinned message");
-            }
+            if ($unpinChatMessage != true) $chat->sendMessage("there isn't any pinned message");
         }
 
 
         // command to unpin all the pinned messages in a group/supergroup
-        if ($text === "/unpinAllMessages") {
-            if ($chat->pinned_message == false) {
-                $chat->sendMessage("no pinned messages");
-            } else {
-                $unpinAllChatMessage = $Bot->unpinAllChatMessages([
+        if ($text === "/unpinAllMessages") 
+        {
+            if ($chat->pinned_message == false) $chat->sendMessage("no pinned messages");
+            else 
+            {
+                $Bot->unpinAllChatMessages
+                ([
                     "chat_id" => $chatid
                 ]);
             }
@@ -114,8 +110,10 @@ try {
 
 
         // command to mute the user in reply
-        if ($text == "/muteUser") {
-            $restrict = $Bot->restrictChatMember([
+        if ($text == "/muteUser") 
+        {
+            $restrict = $Bot->restrictChatMember
+            ([
                 "chat_id" => $chatid,
                 "user_id" => $useridReply,
                 "can_send_messages" => 0
@@ -125,8 +123,10 @@ try {
 
 
         // command to unmute the user in reply
-        if ($text == "/unmuteUser") {
-            $Bot->restrictChatMember([
+        if ($text == "/unmuteUser") 
+        {
+            $Bot->restrictChatMember
+            ([
                 "chat_id" => $chatid,
                 "user_id" => $useridReply,
                 "can_send_messages" => 1
@@ -136,11 +136,13 @@ try {
 
         // command to make admin a user
         // example: "/promoteMember 1 1 1 1 1" 
-        if ((strpos($text, "/promoteMember") === 0) and ($replyToMessage !== null and $usernameReply !== null)) {
+        if ((strpos($text, "/promoteMember") === 0) and ($replyToMessage !== null and $usernameReply !== null))
+        {
             $ai = str_replace("admin ", "", $text);
             $ai = explode(" ", $text);
 
-            $Bot->promoteChatMember([
+            $Bot->promoteChatMember
+            ([
                 "chat_id" => $chatid,
                 "user_id" => $useridReply,
                 "can_change_info" => boolval($ai[0]),
@@ -154,8 +156,10 @@ try {
 
 
         // command to demote an admin to user
-        if ((strpos($text, "/demoteMember") === 0) and ($replyToMessage !== null and $usernameReply !== null)) {
-            $Bot->promoteChatMember([
+        if ((strpos($text, "/demoteMember") === 0) and ($replyToMessage !== null and $usernameReply !== null)) 
+        {
+            $Bot->promoteChatMember
+            ([
                 "chat_id" => $chatid,
                 "user_id" => $useridReply,
                 "can_change_info" => false,
@@ -168,13 +172,16 @@ try {
 
 
         // command to list admins of the chat
-        if ($text == "/listAdmins") {
-            $chatAdmins = $Bot->getChatAdministrators([
+        if ($text == "/listAdmins")
+        {
+            $chatAdmins = $Bot->getChatAdministrators
+            ([
                 "chat_id" => $chatid
             ]);
 
             $adminsar = [];
-            foreach ($chatAdmins as $chatAdmin) {
+            foreach ($chatAdmins as $chatAdmin) 
+            {
                 if ($chatAdmin === '[ChatMember]') continue;
                 
                 $adminsar[] = "<a href='t.me/{$chatAdmin->user->username}'>@" . ($chatAdmin->user->username) . "</a>\n";
@@ -185,27 +192,27 @@ try {
 
 
         // command to check if a user is an admin
-        if ($text == "/isAdmins?") {
-            $chatMember = $Bot->getChatMember([
+        if ($text == "/isAdmins?") 
+        {
+            $chatMember = $Bot->getChatMember
+            ([
                 "chat_id" => $chatid,
                 "user_id" => $useridReply
             ]);
 
             $isAdmin = in_array($chatMember->status, ["administrator", "creator"]);
 
-            if ($isAdmin) {
-                $chat->sendMessage("@" . $usernameReply . " is an admin");
-            } else {
-                if (!in_array($chatMember->status, ["administrator", "creator"])) {
-
-                    $chat->sendMessage("@" . $usernameReply . " isn't an admin");
-                }
+            if ($isAdmin) $chat->sendMessage("@" . $usernameReply . " is an admin");
+            else 
+            {
+                if (!in_array($chatMember->status, ["administrator", "creator"])) $chat->sendMessage("@" . $usernameReply . " isn't an admin");
             }
         }
 
 
         // check lag (?)
-        if ($text === "/checkLag") {
+        if ($text === "/checkLag") 
+        {
             $start = hrtime(true);
             $lagMex = $chat->sendMessage("lag...");
             $time_elapsed_secs = hrtime(true) - $start;
@@ -214,7 +221,8 @@ try {
 
 
         // command to get user_id of user in reply
-        if ($text === "/userId") {
+        if ($text === "/userId") 
+        {
             $chat->sendMessage("@$usernameReply id: <code>$useridReply</code>");
         }
 
@@ -223,7 +231,8 @@ try {
         // use this if your VM is based on Linux
         // use if hosting provider allows to do the "shell_exec()" command
         // usage: /command command, example: "/command ls" 
-        if (stripos($text, "/command") === 0) {
+        if (stripos($text, "/command") === 0) 
+        {
             $explode = explode(" ", $text);
             unset($explode[0]);
             $command = implode(" ", $explode);
@@ -234,7 +243,8 @@ try {
     }
 
 
-} catch (Throwable $e) {
+} catch (Throwable $e) 
+{
     file_put_contents("logs.txt", $e . "\n", FILE_APPEND);
     $chat->sendMessage("<b>error code: " . $e->getCode() . "</b>\n\n<b>description: </b>$e");
 }
